third_party/blink/web_tests/external/wpt/css/css-box/margin-trim/computed-margin-values/flexbox-row-multi-line-inline-start-expected.txt

This is a testharness.js-based test.
[FAIL] flexbox > item 1
  assert_equals: \n<item data-expected-margin-left="0" data-offset-x="8"></item>\noffsetLeft expected 8 but got 18
[FAIL] flexbox > item 2
  assert_equals: \n<item data-expected-margin-left="-10" data-offset-x="48"></item>\noffsetLeft expected 48 but got 58
[FAIL] flexbox > item 3
  assert_equals: \n<item data-expected-margin-left="0" data-offset-x="8"></item>\noffsetLeft expected 8 but got 63
[FAIL] flexbox > item 4
  assert_equals: \n<item data-expected-margin-left="10" data-offset-x="68"></item>\noffsetLeft expected 68 but got 18
Harness: the test ran to completion.
